add rolebase permissions on product endpoints

+++ src/modules/product/controller.ts
@@ -4,6 +4,7 @@ import createHttpError from 'http-errors';
 import Product from './model.js';
 import storage from '../../services/storage.js';
 import mongoose from 'mongoose';
+import { UserRole, type AuthInterface } from '../../types/index.js';
 
 export const createProduct = async (
     req: Request,
@@ -16,6 +17,7 @@ export const createProduct = async (
         if (!valid.isEmpty()) {
             return next(createHttpError(400, valid.array()[0]?.msg));
         }
+        const request = req as AuthInterface;
 
         // values
         const {
@@ -30,31 +32,45 @@ export const createProduct = async (
 
         const priceConfig = JSON.parse(jsonPriceConfigFormat);
 
-        const image = req.file;
-        if (!mongoose.Types.ObjectId.isValid(categoryId as string)) {
-            return next(createHttpError(400, 'Category Id should be valid'));
+        if (
+            request.auth.role === UserRole.ADMIN ||
+            (request.auth.tenant && request.auth.tenant.id === tenantId)
+        ) {
+            const image = req.file;
+            if (!mongoose.Types.ObjectId.isValid(categoryId as string)) {
+                return next(
+                    createHttpError(400, 'Category Id should be valid')
+                );
+            }
+            if (!image) {
+                return next(createHttpError(400, 'Image is required'));
+            }
+
+            const imageUrl = await storage.upload(
+                image,
+                image.originalname as string
+            );
+
+            const product = await Product.create({
+                name,
+                description,
+                attributes,
+                categoryId,
+                isPublished,
+                priceConfig,
+                imageUrl,
+                tenantId,
+            });
+
+            return res.status(201).json({ success: true, product });
+        } else {
+            return next(
+                createHttpError(
+                    400,
+                    "You don't have permission for this action"
+                )
+            );
         }
-        if (!image) {
-            return next(createHttpError(400, 'Image is required'));
-        }
-
-        const imageUrl = await storage.upload(
-            image,
-            image.originalname as string
-        );
-
-        const product = await Product.create({
-            name,
-            description,
-            attributes,
-            categoryId,
-            isPublished,
-            priceConfig,
-            imageUrl,
-            tenantId,
-        });
-
-        return res.status(201).json({ success: true, product });
     } catch (error) {
         return next(new Error(error as string));
     }
@@ -86,43 +102,74 @@ export const updateProduct = async (
             ? JSON.parse(jsonPriceConfigFormat)
             : undefined;
         const { id } = req.params;
+        const request = req as AuthInterface;
 
         if (!mongoose.Types.ObjectId.isValid(id as string)) {
             return next(createHttpError(400, 'Category Id is not valid'));
         }
 
-        const image = req.file;
-        let imageUrl;
+        const tenant = await Product.findById(id);
 
-        if (image) {
-            const currentProduct = await Product.findById(id);
-
-            const ImageKeyId = currentProduct?.imageUrl;
-            if (ImageKeyId) {
-                await storage.delete(ImageKeyId as string);
-            }
-
-            imageUrl = await storage.upload(
-                image,
-                image.originalname as string
+        if (!tenant) {
+            return next(
+                createHttpError(400, 'No product found with matching id')
             );
         }
 
-        const updatedProduct = await Product.findByIdAndUpdate(id, {
-            ...(priceConfig && { priceConfig }),
-            ...(isPublished && { isPublished }),
-            ...(categoryId && { categoryId }),
-            ...(description && { description }),
-            ...(name && { name }),
-            ...(imageUrl && { imageUrl }),
-            ...(attributes && { attributes }),
-        });
+        if (
+            request.auth.role === UserRole.ADMIN ||
+            (request.auth.tenant && request.auth.tenant.id === tenant.tenantId)
+        ) {
+            const image = req.file;
+            let imageUrl;
 
-        return res.status(200).json({ success: true, product: updatedProduct });
+            if (image) {
+                const currentProduct = await Product.findById(id);
+
+                const ImageKeyId = currentProduct?.imageUrl;
+                if (ImageKeyId) {
+                    await storage.delete(ImageKeyId as string);
+                }
+
+                imageUrl = await storage.upload(
+                    image,
+                    image.originalname as string
+                );
+            }
+
+            const updatedProduct = await Product.findByIdAndUpdate(
+                id,
+                {
+                    ...(priceConfig && { priceConfig }),
+                    ...(isPublished !== undefined && { isPublished }),
+                    ...(categoryId && { categoryId }),
+                    ...(description && { description }),
+                    ...(name && { name }),
+                    ...(imageUrl && { imageUrl }),
+                    ...(attributes && { attributes }),
+                },
+                {
+                    new: true,
+                    runValidators: true,
+                }
+            );
+
+            return res
+                .status(200)
+                .json({ success: true, product: updatedProduct });
+        } else {
+            return next(
+                createHttpError(
+                    403,
+                    "You don't have permission for this action"
+                )
+            );
+        }
     } catch (error) {
         return next(new Error(error as string));
     }
 };
+
 export const getProduct = async (
     req: Request,
     res: Response,
@@ -136,10 +183,14 @@ export const getProduct = async (
         }
 
         if (!mongoose.Types.ObjectId.isValid(id as string)) {
-            return next(createHttpError(400, 'Category Id is not valid'));
+            return next(createHttpError(400, 'Product Id is not valid'));
         }
 
         const product = await Product.findById(id);
+
+        if (!product) {
+            return next(createHttpError(404, 'Product not found'));
+        }
 
         return res.status(200).json({ success: true, product });
     } catch (error) {
@@ -154,6 +205,7 @@ export const deleteProduct = async (
 ) => {
     try {
         const { id } = req.params;
+        const request = req as AuthInterface;
 
         if (!id) {
             return next(createHttpError(400, 'Product Id is required'));
@@ -163,17 +215,31 @@ export const deleteProduct = async (
             return next(createHttpError(400, 'Category Id is not valid'));
         }
 
-        const currentProduct = await Product.findById(id);
+        const product = await Product.findById(id);
 
-        if(!currentProduct){
-           return next(createHttpError(404, "No related product found to delete"))
+        if (!product) {
+            return next(
+                createHttpError(400, 'No product found with matching id')
+            );
         }
-        
-        await storage.delete(currentProduct.imageUrl)
 
-        const product = await Product.findByIdAndDelete(id);
+        if (
+            request.auth.role === UserRole.ADMIN ||
+            (request.auth.tenant && request.auth.tenant.id === product.tenantId)
+        ) {
+            await storage.delete(product.imageUrl);
 
-        return res.status(200).json({ success: true, product });
+            const deletedProduct = await Product.findByIdAndDelete(id);
+
+            return res.status(200).json({ success: true, deletedProduct });
+        } else {
+            return next(
+                createHttpError(
+                    403,
+                    "You don't have permission for this action"
+                )
+            );
+        }
     } catch (error) {
         return next(new Error(error as string));
     }
